Refactor ColdStorage: Remove trait, consolidate Redb impl, delete InMemory impl
